Renaming common module to core and refactoring package name for repo implementation
