+++ powertools-idempotency/powertools-idempotency-core/src/test/resources/apigw_event.json
